fusermount: Fix option escaping
Before this patch no commas were allowed in any fusermount option string.
But it is in fact possible to use commas in fusermount options: we only
need to escape them [1].

Now go-fuse users can use commas in fusermount options.

Related libfuse commits are [1] and [2].

This patch also adds a test which ensures neither go-fuse nor fusermount
crash if we pass options with commas.

---

[1] https://github.com/libfuse/libfuse/commit/555d6b504308eac6b976321ce938ee4bec62c354
[2] https://github.com/libfuse/libfuse/commit/5c094ac0150ebfef6a2c9c2c9d1c545a90ff4e96

Change-Id: Iadf8283e630dc1fb884d047d8765eeaaca90eabd