fuse/test: Fix TestLargeDirRead
This test tries to create many files with random names and then asserts
that readdir returns so many files that it created. But the test
had a bug in that two random file names could be the same
name but accounted twice. This leads to assertion in the end that the
number of direntries read via readdir is not as expected.

Here is how two random names could turn out to be the same:

name 1: "file3" + "0" generated by randomLengthString
name 2: "file30" + "" generated by randomLengthString

-> Fix it by constructing random names guaranteed different.

Fixes: https://github.com/hanwen/go-fuse/issues/472
Change-Id: I776866bd728479da7324878d41dc71e504efd229