fs: increase the limit for TestFileFdLeak
The RELEASE opcode is asynchronous, so the next open() call may
execute before the RELEASE for the preceding one has completed. In
case of GOMAXPROCS=1, this happens often.

Increase the limit to 15 to reduce the chance of this causing a test
failure. The test runs 100 times, so this should still catch logic
errors with deallocating file handles.

Fixes #477.

Change-Id: I12dbd96b2a6082e0f4a386d35983e6efcc596aef