fs: bridge: Add missing Lseek whence checks
The man page for lseek(2) says that the filesystem should
return an ENXIO when the offset is greater than the file
size on either SEEK_DATA or SEEK_HOLE or if the offset is
set to the end of the file with SEEK_DATA (where there is
an implicit hole). Adding these checks to the current Lseek
implementation.

Change-Id: I8db0913a9e37209e1681ffd11a9afb7f39810132