fs: new directory API
The new API follows the pattern of the other file API:

InodeEmbedder can implement a OpendirHandle method, which returns a
FileHandle. Directories can implement the following APIs,

 * FileSyncdirer
 * FileReaddirenter
 * FileReleasedirer
 * FileSeekdirer

along with the opened directory, FOPEN_* flags may be returned. In a
follow-up change, we'll demonstrate how to use FOPEN_CACHE_DIR for
directory caching.

This change is backward compatible in a compilation sense, but support
for seeking in DirStreams has been removed: it was incorrect (it made
the assumption that the dir offset was monotonically increasing),
inefficient and complicated.

Change-Id: I662713321f0f812e92e4059ee11e8b1427c6aa0f