fs: fix negative lookup cache
For returning a negative lookup result, LOOKUP must return a 0
errno. If not, the return value (including its EntryOut with entry
timeout set) is truncated in the lower-level machinery.

This was not obvious from the debug output, as that still prints
ENOENT + EntryOut. Fix this in a separate commit.

Add test.

Change-Id: I4f392301e937b38ba116114befab35cde4e3044d