vhostuser: vhost server for virtioFS
This is experimental support serving FUSE directly to a VM using
virtiofs.

This originally was built end of 2024
(https://github.com/hanwen/go-fuse/tree/virtiofs), by following QEMU's
control flow for virtiofsd, and languished since.

With the help of Claude, I resolved most of the outstanding issues.

This is an internal library, because I barely know what I am doing,
and do not want the overhead of stability/compatibility for external
users

This also bumps minimum Go version to 1.21.