fuse: fix memory leak for read buffers
In commit 9c6a577 ("fuse: drop server.allocOut"), I intended to move
where requestAlloc.bufferPoolOutputBuf was set, but I actually removed
the line.

This effectively disables the sync.Pool used for shipping large data
(ie. reads) back to the kernel. This was visible in the benchmark
outputs, but that requires close inspection to notice the problem. 

Fix this, and add a test that catches this problem in the future.

Thanks to https://github.com/darthShadow for diagnosing and reporting
this problem.

Fixes https://github.com/hanwen/go-fuse/issues/542

Change-Id: I40e17274c8e8f2b12e46008661c324c68fc81335