fuse: enable id-mapped mount on supported kernels
id-mapped mount for fuse was introduced in Linux kernel v6.12.
This patch added a config option in go-fuse to enable the
id-mapped mount feature on supported kernels.

Based on https://lwn.net/Articles/985803/, id-mapped mount
for fuse requires the filesystem mounted with the "default_permissions"
parameter and internally trust the kernel to perform the
UID/GID-based checks.

For id-mapped mounts, fuse will send mapped UID/GID in requests
that create new inodes, and -1 for the others.

An unit test was also written to validate the feature via
the updated loopback implementation. The test requires root
privilege to run due to the open_tree syscall.

Change-Id: I6a93a0cd2109a03c5bb54946fba00c535d9d3d21