fuse: enable id-mapped mount on supported kernels
The id-mapped mount feature for FUSE was introduced in Linux v6.12.
This patch added a config option to enable the feature on supported kernels.

Using id-mapped mount for FUSE requires the filesystem mounted with the "default_permissions" parameter and must trust the kernel to perform UID/GID-based checks correctly.

For id-mapped mounts, FUSE will send mapped UID/GID in requests
that create new inodes, and -1 for the others.

More information in https://lwn.net/Articles/985803/ .

A unit test was also written to validate the feature. The test
requires root privilege to run due to the open_tree syscall.

Change-Id: I6a93a0cd2109a03c5bb54946fba00c535d9d3d21