fuse: fix NewLoopbackDirStream on darwin
The rawBridge expects the returned FileHandle from OpendirHandle to be a FileReaddirenter.
However, on Darwin, the NewLoopbackDirStream used in LoopbackNode.OpendirHandle returns a dirArray,
which is not a FileReaddirenter. Consequently, the fs.NewLoopbackRoot function does not work on Darwin.

To address this issue, this commit implements FileReaddirenter/FileReleasedirer for dirArray.

Signed-off-by: Rueian <rueiancsie@gmail.com>