fs: add NewLoopbackDirStreamFd
This allows use of loopbackDirStream in a
symlink-race-safe way.

A symlink-race attack could allow an attacker to
escape from the directory tree and read something
like /etc.

This may or may not be a concern for a filesystem
implementation, but it is for gocryptfs,
and gocryptfs would like to use loopbackDirStream via
NewLoopbackDirStreamFd.

Change-Id: I00792c697e2ef9b9be6ed1cc251e29b886a31892