fuse: fix mount failure for 7.22 and earlier
`example/hello` fails to mount on my legacy Linux 3.10, with the following
error message:

    writer: Write/Writev failed, err: 22=invalid argument. opcode: INIT
    Mount fail: init: 22=invalid argument

`strace` shows that the failure occurs when replying the INIT command.

    write(9,
          "P\0\0\0\0\0\0\0\1\0\0\0\0\0\0\0\7\0\0\0\26\0\0\0\0\20\0\0!0\0@"...,
          80) = -1 EINVAL (Invalid argument)

The length of InitOut (without common header) should be 24 in fuse
protocol 7.22, rather than 64.

In protocol-server.go#L70 an attempt was made to address this, but the
fixup was overwritten in serializeHeader(), so do it there instead.

Change-Id: I8211e13ca8fe06460b11fcdbfc340a71205c0417