fuse: change to create a fixed number of loop() goroutines
Change to create a fixed number of loop() goroutines. And make
`MaxLoopNum` configurable.

Current `maxReaders` not only controls the number of FUSE readers, but
also controls the number of loop() goroutines running. As it will keep
creating loop() goroutings, after it creates the `maxReaders` number of
loop() goroutings the redundant loop() goroutings will just exit.
Printings show that most of the creating loop() goroutines will exit for
test BenchmarkGoFuseMemoryRead-112(runs 30s):
10:19:21.754783 Serve(): First loop created, loopCreateNum: 1
10:19:55.568646 Exit... maxReaders: 16, reqReaders: 17, reqNum: 8870103,
	loopCreateNum: 110081, loopExitNum: 110064, curLoopNum: 17

This might lead to some overhead of maintaining the redundant loop()
goroutines for the Go runtime, such as goroutine creation, scheduling,
destruction, and garbage collection.
Tests show that this overhead is non-negligible for Arm64 machines(but
not x86):
For Go-fuse test BenchmarkGoFuseFDWrite and downstream Juicefs sequential
write/read tests with this patch, the max bandwidth has a 1.25x+ gain.
Specifically, the Juicefs sequential write max bandwidth increases from
14.9GiB/s to 18.7GiB/s.
For more details about test results, please refer to[1].

So why not just create the required number of loop() goroutines and let
them live on to serve the FUSE requests like Libfuse[2]. This makes the
code logic clear and simple.

[1]
https://github.com/hanwen/go-fuse/issues/565#issuecomment-2961209380
[2]
https://github.com/libfuse/libfuse/blob/cea9b33acc13e7683e3646fca04953d14e3211b7/lib/fuse_loop_mt.c#L174-L175

Related-To: https://github.com/hanwen/go-fuse/issues/565
Change-Id: I8582abe9ddbd67a4bd304d527530f054cef796ad
Signed-off-by: Xinliang Liu <xinliang.liu@linaro.org>