benchmark: add a performance benchmark based on fio
This is more representative of production use (The previous read
benchmark does nothing, simply returning the input buffer unchanged).
For a pure memory read (mode=read), the numbers are similar though, as
MemRegularFile also simply returns an existing memory block to the
kernel.

t480 i5-8350U, CPU pinned @ 2ghz:

FIO:

    Run status group 0 (all jobs):
       READ: bw=1796MiB/s (1884MB/s), 1796MiB/s-1796MiB/s (1884MB/s-1884MB/s), io=1024MiB (1074MB), run=570-570msec

Go

    BenchmarkGoFuseMemoryRead
    20313	     52985 ns/op	2473.75 MB/s	      79 B/op     3 allocs/op

This is modeled as a test for now. For benchmarking, there are various
interesting things to consider (number of threads, blocksize, mode,
etc.), but they cannot necessarily be expressed as golang benchmarking
options. This is left to consider in a future change.

Change-Id: Id0000000ea049366a06efc8c03969e6ffb095914