fuse: add `EnableWritebackCaching` option to control writeback cache mode
This commit adds a new `EnableWritebackCaching` option that allows users
to set the CAP_WRITEBACK_CACHE when mounting the filesystem. This option
changes the way in which write requests are processed reach the
filesystem server: instead of each write being immediately sent to userspace
as one or more WRITE requests, writes will be satisfied directly from
the page cache. Since this option may have negative repercussions in
networked filesystems, it is disabled by default.

Signed-off-by: Alberto Miranda <alberto.miranda@seqera.io>